Added LevelManager.CenterLevelToZeroPosition, Added Zooming, tested Scaledown of cube each 20 secs, etc
CenterLevelToZeroPosition we now center our cube level by its bounds.

+++ Assets/Scripts/UIManager.cs
@@ -26,7 +26,7 @@ public class UIManager : MonoBehaviour {
 
     public void UpdateScore()
     {
-        scoreText.text = "Score: " + GameManager.Instance.currentScore;
+        scoreText.text = "Score: " + (int)GameManager.Instance.currentScore;
     }
 
     public void UpdateTime()
